Added jest, moved main logic to src, wrote tests for strictMode
